Work Done In Task 3
Random Forest Classifier Applied with Feature Engineering -> Code 1

Updated Random Forest Classifier Applied with Feature Engineering -> Code 2

Why did you choose the evaluation metrics that you used? Please elaborate on your choices.

Document the advantages and disadvantages of using the Random Forest for this use case.

Do you think that the model performance is satisfactory? Give justification for your answer.

Bonus) - Relate the model performance to the client's financial performance with the introduction of the discount proposition. How much money could a client save with the use of the model? What assumptions did you make to come to this conclusion?